% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphutils-package.R
\docType{package}
\name{alphutils-package}
\alias{alphutils}
\alias{alphutils-package}
\title{alphutils: Utilities for Alephium}
\description{
A bunch of utlity functions to retrieve, modify and analyse Alephium related data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dm807cam/aleputils}
  \item Report bugs at \url{https://github.com/dm807cam/aleputils/issues}
}

}
\author{
\strong{Maintainer}: Dennis Mayk \email{dm807@cantab.ac.uk} (\href{https://orcid.org/0000-0001-6762-2291}{ORCID})

}
\keyword{internal}
