% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_address.R
\name{api_address_locked_balance}
\alias{api_address_locked_balance}
\title{Fetch number of locked ALPH}
\usage{
api_address_locked_balance(
  address,
  endpoint = explorer_endpoint_select(),
  req_timeout = 300
)
}
\arguments{
\item{address}{\strong{\link{character}}: A valid Alephium address}

\item{endpoint}{\strong{\link{character}}: URL to the explorer API}

\item{req_timeout}{\strong{\link{integer}}: Number of seconds until request times out}
}
\value{
\strong{\link{double}}: Number of locked $ALPH
}
\description{
Fetch the number of locked $ALPH in a specific address.
}
