% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_address.R
\name{api_address_richlist}
\alias{api_address_richlist}
\title{Fetch rich list}
\usage{
api_address_richlist(
  min_balance = 1,
  endpoint = explorer_endpoint_select(),
  req_timeout = 300,
  req_sleep = 1
)
}
\arguments{
\item{min_balance}{\strong{\link{double}}: Min balance of an address to be included in the output}

\item{endpoint}{\strong{\link{character}}: URL to the explorer API}

\item{req_timeout}{\strong{\link{integer}}: Number of seconds until request times out}

\item{req_sleep}{\strong{\link{integer}}: Number of seconds between requests. Important to deal with rate limits.}
}
\value{
\strong{\link{data.table}}: Rich list
}
\description{
Fetch addresses from explorer rich list.
This function iterates through the rich list and fetches all
addresses with their respective balances until no more addresses
with a balance >= \code{min_balance} can be found.
}
