% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{api_GET}
\alias{api_GET}
\title{GET request}
\usage{
api_GET(url, req_timeout)
}
\arguments{
\item{url}{\strong{\link{character}}: URL to the API endpoint}

\item{req_timeout}{\strong{\link{integer}}: Number of seconds until request times out}
}
\value{
\strong{\link{list}}: API response content
}
\description{
GET request
}
