% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.R
\name{api_GET_with_retry}
\alias{api_GET_with_retry}
\title{GET request with retry}
\usage{
api_GET_with_retry(url, req_timeout, max_retries = 5)
}
\arguments{
\item{url}{\strong{\link{character}}: URL to the API endpoint}

\item{req_timeout}{\strong{\link{integer}}: Number of seconds until request times out}

\item{max_retries}{\strong{\link{integer}}: Number of retries before the function fails}
}
\value{
\strong{\link{list}}: API response content
}
\description{
Wrapper function to \code{api_GET()} that retries \code{max_retries} times if a request failed.
}
