% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_utils.R
\name{convert_pi_to_alph}
\alias{convert_pi_to_alph}
\title{Convert Pi to ALPH}
\usage{
convert_pi_to_alph(pi)
}
\arguments{
\item{pi}{\strong{\link{double}}: Amount of Pi converted to $ALPH}
}
\value{
\strong{\link{double}}: $ALPH
}
\description{
Convert Pi to ALPH
}
