% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoint.R
\name{explorer_endpoint_select}
\alias{explorer_endpoint_select}
\title{Explorer API URL}
\usage{
explorer_endpoint_select(ii = 1)
}
\arguments{
\item{ii}{\strong{\link{integer}}: Index of from internal URL vector to be returned. This is useful if multiple URL ought to be tried sequentially.}
}
\value{
\strong{\link{character}}: Explorer API URL
}
\description{
Supplies a number of different explorer API endpoints.
}
