% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_utils.R
\name{true_random}
\alias{true_random}
\title{Fetch random number(s)}
\usage{
true_random(n = 1, min = 1, max = 100, max_retries = 5)
}
\arguments{
\item{n}{\strong{\link{integer}}: Length of random vector}

\item{min}{\strong{\link{integer}}: Min number to consider by random.org}

\item{max}{\strong{\link{integer}}: Max number to consider by random.org}

\item{max_retries}{\strong{\link{integer}}: Max number of retries if random.org returns an invalid repsonse}
}
\value{
\strong{\link{integer}}: Vector
}
\description{
Fetch truly random number(s) from random.org
}
